requirements.txt
Flask>=3.0,<4
python-dotenv>=1.0,<2
requests>=2.31,<3
